Add ResNetFeature
