+++ jsk_recognition_utils/python/jsk_recognition_utils/chainermodels/__init__.py
@@ -20,5 +20,8 @@ VGG_CNN_M_1024 = vgg_cnn_m_1024.VGG_CNN_M_1024
 
 # ResNet
 ResNet50 = resnet.ResNet50
+ResNet50Feature = resnet.ResNet50Feature
 ResNet101 = resnet.ResNet101
+ResNet101Feature = resnet.ResNet101Feature
 ResNet152 = resnet.ResNet152
+ResNet152Feature = resnet.ResNet152Feature
