+++ jsk_recognition_utils/python/jsk_recognition_utils/chainermodels/resnet/__init__.py
@@ -1,3 +1,6 @@
 from .resnet50 import ResNet50  # NOQA
+from .resnet50 import ResNet50Feature  # NOQA
 from .resnet101 import ResNet101  # NOQA
+from .resnet101 import ResNet101Feature  # NOQA
 from .resnet152 import ResNet152  # NOQA
+from .resnet152 import ResNet152Feature  # NOQA
